+++ CHANGELOG.md
@@ -1,11 +1,42 @@
 # Change Log
 
+## [3.3.0](https://github.com/EddyVerbruggen/nativescript-mapbox/tree/3.3.0) (2018-01-20)
+[Full Changelog](https://github.com/EddyVerbruggen/nativescript-mapbox/compare/3.2.0...3.3.0)
+
+**Fixed bugs:**
+
+- Console output query [\#177](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/177)
+
+**Implemented enhancements:**
+
+- listOfflineRegions fails on Android if no map has been shown [\#172](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/172)
+- How to get lat long of userlocation from mapbox [\#180](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/180)
+- Failed to bind EAGLDrawable [\#196](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/196)
+
+
+## [3.2.0](https://github.com/EddyVerbruggen/nativescript-mapbox/tree/3.2.0) (2018-01-13)
+[Full Changelog](https://github.com/EddyVerbruggen/nativescript-mapbox/compare/3.1.3...3.2.0)
+
+**Fixed bugs:**
+
+- Dragging Map [\#175](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/175)
+- Don't handle permissions other than our own [\#194](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/194)
+- Don't fire the "mapready" event when the style changes [\#195](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/195)
+
+**Implemented enhancements:**
+
+- Is there a way to capture when the map is scrolled? [\136](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/136)
+- New Listeners [\#183](https://github.com/EddyVerbruggen/nativescript-mapbox/pull/183)
+- Calling selectAnnotation() from native map view throws "TypeError: nv.selectAnnotation is not a function" [\#186](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/186) 
+- Bump Mapbox SDK versions [\#193](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/193)
+
+
 ## [3.1.3](https://github.com/EddyVerbruggen/nativescript-mapbox/tree/3.1.3) (2017-11-02)
 [Full Changelog](https://github.com/EddyVerbruggen/nativescript-mapbox/compare/3.1.2...3.1.3)
 
 **Fixed bugs:**
 
-- Marker onTap - after navigating back to map [\#125](https://github.com/EddyVerbruggen/nativescript-mapbox/pull/125)
+- Marker onTap - after navigating back to map [\#125](https://github.com/EddyVerbruggen/nativescript-mapbox/issues/125)
 - Markers not cleaned from cache on android [\#182](https://github.com/EddyVerbruggen/nativescript-mapbox/pull/182)